// src/app/interfaces/bill.ts
export interface Bill {
  comercio_codigo: number;
  comercio_nombre: string;
  comercio_nit: string;
  comercio_direccion: string;
  Trans_codigo: string;
  Trans_medio_pago: number;
  Trans_estado: number;
  Trans_total: number;
  Trans_fecha: string;
  Trans_concepto: string;
  usuario_identificacion: string;
  usuario_nombre: string;
  usuario_email: string;
}